staticTmpDir: enable forcing tmpDir to a static path
Added staticTmpDir option to Config.
This will force tmpDir to be created at a static path.
Tmp dir will be autodeleted at exit, maybe that
`staticTmpDirLeaveAfterError` was defined, then we'll show a warning that
it needs to be manually cleaned up.

This helps with using stryker4s with Bazel, as Bazel uses the path to the
workspace (folder) to determine where to store the local cache (hashes
the path to create the sandbox's name).
Enabled not autoremoving the folder to be able to debug issues (as Bazel
support is early-stage).